requirements.txt
python-nmap==0.6.1
paramiko==2.6.0
ansible==2.7.8
netifaces==0.10.9
netaddr==0.7.19
